tst_qmodbuspdu: Fix compiler warning
tst_qmodbuspdu.cpp(859): warning C4309: 'argument': truncation of constant value

+++ tests/auto/qmodbuspdu/tst_qmodbuspdu.cpp
@@ -856,7 +856,7 @@ private slots:
         QCOMPARE(QModbusRequest::calculateDataSize(wfrRequest), 1 + longData.size());
 
         longData = QByteArray(4 + 128, ' ');
-        longData[4] = 128;
+        longData[4] = quint8(128);
         const QModbusRequest wmcRequest(QModbusPdu::WriteMultipleCoils, longData);
         QCOMPARE(QModbusRequest::calculateDataSize(wmcRequest), 1 + longData.size());
